requirements.txt
Click==7.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==0.16.1
pyjwt
email_validator

# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
psycopg2


# ----------------------
# Production Server
# ----------------------
gunicorn